Add `textDocument/definition` support between multi files and multi modules.
The solution is a bit brute force, as we always search in all semanticdb
file for the given symbol, but it do the job. We could optimize by
being smarter when listing all files, trying to search first packages
that match the request for example.

Some cleaning needs also to be done to have proper types and use less
`IO`. We should also use some monad transformers....

Note: naming sucks...