Try an optimistic search to find the definition
Most of the time, the good practice is to have the same file name as
your class/object/trait, so from the symbol we extract the
possible name of the file and try to search in it the symbol.
If it's not there, then we search in all project files as previously.

Note: I know the chaining of `case...Just...Nothing` isn't nice, but
it's currently easier for me :)