Unify `JSONRPC` messages handling
Prior to this commit we had to ways of parsing and handling `JSONRPC`
messages, one for the `sbt` messages and another one for the client
communication.

The solution is not yet perfect as the main function is full of side
effect.

Next changes will make use of a proper `State` monad, without `IO` everywhere.